requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.10
matplotlib>=3.7
arch>=6.0
yfinance>=0.2.18
pyarrow>=12.0
pyyaml>=6.0
pytest>=8.0
optuna>=3.0
